target/arm: Fix PC for AArch64 (including Morello) WFI trapping
Upstream's code is env->pc -= insn_len, like the AArch32 case, which is
doing a binary subtraction on the promoted (to target_ulong, i.e.
64-bit) insn_len. However, by using a unary minus downstream, we instead
negate the 32-bit value prior to promoting (again to target_ulong, the
type for the callee's argument) and thus, given insn_len is unsigned,
that means we zero-extend rather than sign-extend. This has the result
of causing PC to be 2^32 more than it should be.