Avoid using cs_base for CHERI
Commit a378206a "target/arm: Move mode specific TB flags to tb->cs_base"
now uses this field for additional TB flags as it was unused in the Arm
target before. We were using this field downstream to store the base of
PCC (which IMO is appropriate considering how it is used in other targets).

While we could reuse cs_base to store the CHERI flags now, this commit uses
a simpler approach of adding another field for pcc_base to avoid future
merge conflicts. This can be cleaned up once we have caught up with latest
QEMU.