cheri-tagmem: relax an assertion on memory region size
We previously asserted that the memory region is exactly a multiple of the
internal tagblock size, but this is not true for some of the unit tests.
For example, the RISC-V tests create a machine with 16KB of RAM, but the
tag block size is 64KB which would cause an assertion here.
Relax the assertion to a warning (although I believe this case is actually
handled correctly). In the future we could drop this warning.