riscv: Add the modesw.{cap,int} instructions
This instruction is needed in order to support a subset of ISAv9 that is
compatible with the upcoming CHERI-RISC-V standard. Code was taken from
the CHERI-Alliance repository and adapted for the current dev branch
(without support for the older 0.90 version of the standard).

It might be possible to optimize this instruction more without dropping
out of the current TCG translation block, but that would require
updating PCC's capmode flag lazily.

Co-authored-by: Paul Buxton <paul.buxton@codasip.com>
Co-authored-by: Martin Kaiser <martin.kaiser@codasip.com>